+Deleting Characters now works Correctly =Changed ID types to Long

+++ src/main/java/com/MMOManagement/SpringApp/Repository/Game/Characters/AspektRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Game.Characters;
 import com.MMOManagement.SpringApp.Model.Game.Characters.Aspekt;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AspektRepository extends JpaRepository<Aspekt, Integer> {
+public interface AspektRepository extends JpaRepository<Aspekt, Long> {
 }
