+++ src/main/java/com/MMOManagement/SpringApp/Repository/Game/Characters/PostacFizycznaRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Game.Characters;
 import com.MMOManagement.SpringApp.Model.Game.Characters.PostacFizyczna;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostacFizycznaRepository extends JpaRepository<PostacFizyczna, Integer> {
+public interface PostacFizycznaRepository extends JpaRepository<PostacFizyczna, Long> {
 }
