+++ src/main/java/com/MMOManagement/SpringApp/Repository/Game/Characters/PostacMagicznaRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Game.Characters;
 import com.MMOManagement.SpringApp.Model.Game.Characters.PostacMagiczna;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostacMagicznaRepository extends JpaRepository<PostacMagiczna, Integer> {
+public interface PostacMagicznaRepository extends JpaRepository<PostacMagiczna, Long> {
 }
