+++ src/main/java/com/MMOManagement/SpringApp/Repository/Game/Characters/PostacWsparciaRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Game.Characters;
 import com.MMOManagement.SpringApp.Model.Game.Characters.PostacWsparcia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostacWsparciaRepository  extends JpaRepository<PostacWsparcia, Integer> {
+public interface PostacWsparciaRepository  extends JpaRepository<PostacWsparcia, Long> {
 }
