+++ src/main/java/com/MMOManagement/SpringApp/Repository/Game/Characters/ZaklecieRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Game.Characters;
 import com.MMOManagement.SpringApp.Model.Game.Characters.Zaklecie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ZaklecieRepository  extends JpaRepository<Zaklecie, Integer> {
+public interface ZaklecieRepository  extends JpaRepository<Zaklecie, Long> {
 }
