+++ src/main/java/com/MMOManagement/SpringApp/Repository/Game/Items/BronRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Game.Items;
 import com.MMOManagement.SpringApp.Model.Game.Items.Bron;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BronRepository extends JpaRepository<Bron, Integer> {
+public interface BronRepository extends JpaRepository<Bron, Long> {
 }
