+++ src/main/java/com/MMOManagement/SpringApp/Repository/Game/Items/PrzedmiotwEkwipunkuRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Game.Items;
 import com.MMOManagement.SpringApp.Model.Game.Items.PrzedmiotWEkwipunku;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PrzedmiotwEkwipunkuRepository  extends JpaRepository<PrzedmiotWEkwipunku, Integer> {
+public interface PrzedmiotwEkwipunkuRepository  extends JpaRepository<PrzedmiotWEkwipunku, Long> {
 }
