+++ src/main/java/com/MMOManagement/SpringApp/Repository/Users/Moderation/JezykRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Users.Moderation;
 import com.MMOManagement.SpringApp.Model.Users.Moderation.Jezyk;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JezykRepository extends JpaRepository<Jezyk,Integer> {
+public interface JezykRepository extends JpaRepository<Jezyk,Long> {
 }
