+++ src/main/java/com/MMOManagement/SpringApp/Repository/Users/Moderation/MistrzGryRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Users.Moderation;
 import com.MMOManagement.SpringApp.Model.Users.Moderation.MistrzGry;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MistrzGryRepository  extends JpaRepository<MistrzGry,Integer> {
+public interface MistrzGryRepository  extends JpaRepository<MistrzGry,Long> {
 }
