+++ src/main/java/com/MMOManagement/SpringApp/Repository/Users/Playerbase/GraczRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Users.Playerbase;
 import com.MMOManagement.SpringApp.Model.Users.Playerbase.Gracz;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GraczRepository  extends JpaRepository<Gracz,Integer> {
+public interface GraczRepository  extends JpaRepository<Gracz,Long> {
 }
