+++ src/main/java/com/MMOManagement/SpringApp/Repository/Users/Playerbase/KlanRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Users.Playerbase;
 import com.MMOManagement.SpringApp.Model.Users.Playerbase.Klan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KlanRepository  extends JpaRepository<Klan,Integer> {
+public interface KlanRepository  extends JpaRepository<Klan,Long> {
 }
